+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### unreleased
 * Drop recursive check on option that failed in several scenarios
+* Drop check for base64 encoded strings which did not work properly in al cases
 * Use WP 5.7 color palette for the UI
 
 ### 1.4.1 ###